Modernize type system for 64-bit compatibility
Update primitive integer and floating-point types to modern Cocoa equivalents:

Header files updated:
- SFTPController.h: Update dotflag, method parameters, and return types
- SFTPMainWindow.h: Update return types
- SFTPTableView.h: Update instance variables and method signatures
- SCPController.h: Update instance variables and method parameters
- SCPTransfer.h: Update method signatures and instance variables
- NSString(SSHAdditions).h: Update method parameters
- NSAttributedString-Ellipsis.h: Update method parameters

Type changes:
- int → NSInteger (for better 64-bit compatibility)
- float → CGFloat (for UI measurements and graphics)
- double → CGFloat (where appropriate for UI context)

These changes ensure proper 64-bit compatibility and follow modern
Cocoa conventions while maintaining backward compatibility.

+++ NSString(SSHAdditions).h
@@ -9,7 +9,7 @@
 
 + ( NSDictionary * )unknownHostInfoFromString: ( NSString * )string;
 + ( NSString * )pathFromBaseDir: ( NSString * )base fullPath: ( NSString * )fullpath;
-+ ( NSString * )clockStringFromInteger: ( int )integer;
++ ( NSString * )clockStringFromInteger: ( NSInteger )integer;
 + ( NSString * )pathForExecutable: ( NSString * )executable;
 - ( NSString * )octalRepresentation;
 - ( char )objectTypeFromOctalRepresentation: ( NSString * )octalRep;
